修改ObjectClass.Put的参数类型为DirType

+++ DynamicPatcher/Projects/Extension.Kratos/Utilities/GiftBoxHelper.cs
@@ -30,7 +30,7 @@ namespace Extension.Utilities
                         var occFlags = pCell.Ref.OccupationFlags;
                         pTechno.Ref.Base.OnBridge = pCell.Ref.ContainsBridge();
                         ++Game.IKnowWhatImDoing;
-                        pTechno.Ref.Base.Put(pCell.Ref.GetCoordsWithBridge(), ((short)DirType.E));
+                        pTechno.Ref.Base.Put(pCell.Ref.GetCoordsWithBridge(), DirType.E);
                         --Game.IKnowWhatImDoing;
                         pCell.Ref.OccupationFlags = occFlags;
                         // 单位放到指定的位置
